move System.Json.Text tests to own project
