+++ src/JsonCryption.System.Text.Json.Tests/AcceptanceTests/Arrays.cs
@@ -1,26 +1,25 @@
 ﻿using Shouldly;
-using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json;
 using Xunit;
 
-namespace JsonCryption.Tests.AcceptanceTests
+namespace JsonCryption.System.Text.Json.Tests.AcceptanceTests
 {
-    public class Enumerables
+    public class Arrays
     {
         [Fact]
-        public void String_enumerable_works()
+        public void String_array_works()
         {
             Coordinator.ConfigureDefault("test");
 
             var myStrings = new[] { "some", "strings", "to test" };
-            var foo = new FooStringEnumerable { MyStrings = myStrings };
+            var foo = new FooStringArray { MyStrings = myStrings };
             var json = JsonSerializer.Serialize(foo);
 
             // make sure it's encrypted
             using (var jsonDoc = JsonDocument.Parse(json))
             {
-                var jsonProperty = jsonDoc.RootElement.GetProperty(nameof(FooStringEnumerable.MyStrings));
+                var jsonProperty = jsonDoc.RootElement.GetProperty(nameof(FooStringArray.MyStrings));
                 jsonProperty.ValueKind.ShouldBe(JsonValueKind.Array);
 
                 var jsonProperties = jsonProperty.EnumerateArray().ToList();
@@ -33,29 +32,29 @@ namespace JsonCryption.Tests.AcceptanceTests
             }
 
             // decrypt and check
-            var decrypted = JsonSerializer.Deserialize<FooStringEnumerable>(json);
+            var decrypted = JsonSerializer.Deserialize<FooStringArray>(json);
             decrypted.MyStrings.ShouldBe(myStrings);
         }
 
-        private class FooStringEnumerable
+        private class FooStringArray
         {
             [Encrypt]
-            public IEnumerable<string> MyStrings { get; set; }
+            public string[] MyStrings { get; set; }
         }
 
         [Fact]
-        public void Int_enumerable_works()
+        public void Int_array_works()
         {
             Coordinator.ConfigureDefault("test");
 
             var myInts = new[] { int.MinValue, -1, 0, 1, int.MaxValue };
-            var foo = new FooIntEnumerable { MyInts = myInts };
+            var foo = new FooIntArray { MyInts = myInts };
             var json = JsonSerializer.Serialize(foo);
 
             // make sure it's encrypted
             using (var jsonDoc = JsonDocument.Parse(json))
             {
-                var jsonProperty = jsonDoc.RootElement.GetProperty(nameof(FooIntEnumerable.MyInts));
+                var jsonProperty = jsonDoc.RootElement.GetProperty(nameof(FooIntArray.MyInts));
                 jsonProperty.ValueKind.ShouldBe(JsonValueKind.Array);
 
                 var jsonProperties = jsonProperty.EnumerateArray().ToList();
@@ -68,14 +67,14 @@ namespace JsonCryption.Tests.AcceptanceTests
             }
 
             // decrypt and check
-            var decrypted = JsonSerializer.Deserialize<FooIntEnumerable>(json);
-            decrypted.MyInts.ShouldBe(myInts);;
+            var decrypted = JsonSerializer.Deserialize<FooIntArray>(json);
+            decrypted.MyInts.ShouldBe(myInts);
         }
 
-        private class FooIntEnumerable
+        private class FooIntArray
         {
             [Encrypt]
-            public IEnumerable<int> MyInts { get; set; }
+            public int[] MyInts { get; set; }
         }
     }
 }
