+++ src/JsonCryption.System.Text.Json.Tests/AcceptanceTests/Built_in_types.cs
@@ -3,7 +3,7 @@ using System;
 using System.Text.Json;
 using Xunit;
 
-namespace JsonCryption.Tests.AcceptanceTests
+namespace JsonCryption.System.Text.Json.Tests.AcceptanceTests
 {
     public class Built_in_types
     {
