+++ src/JsonCryption.System.Text.Json.Tests/AcceptanceTests/Enumerables.cs
@@ -1,25 +1,26 @@
 ﻿using Shouldly;
+using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json;
 using Xunit;
 
-namespace JsonCryption.Tests.AcceptanceTests
+namespace JsonCryption.System.Text.Json.Tests.AcceptanceTests
 {
-    public class Arrays
+    public class Enumerables
     {
         [Fact]
-        public void String_array_works()
+        public void String_enumerable_works()
         {
             Coordinator.ConfigureDefault("test");
 
             var myStrings = new[] { "some", "strings", "to test" };
-            var foo = new FooStringArray { MyStrings = myStrings };
+            var foo = new FooStringEnumerable { MyStrings = myStrings };
             var json = JsonSerializer.Serialize(foo);
 
             // make sure it's encrypted
             using (var jsonDoc = JsonDocument.Parse(json))
             {
-                var jsonProperty = jsonDoc.RootElement.GetProperty(nameof(FooStringArray.MyStrings));
+                var jsonProperty = jsonDoc.RootElement.GetProperty(nameof(FooStringEnumerable.MyStrings));
                 jsonProperty.ValueKind.ShouldBe(JsonValueKind.Array);
 
                 var jsonProperties = jsonProperty.EnumerateArray().ToList();
@@ -32,29 +33,29 @@ namespace JsonCryption.Tests.AcceptanceTests
             }
 
             // decrypt and check
-            var decrypted = JsonSerializer.Deserialize<FooStringArray>(json);
+            var decrypted = JsonSerializer.Deserialize<FooStringEnumerable>(json);
             decrypted.MyStrings.ShouldBe(myStrings);
         }
 
-        private class FooStringArray
+        private class FooStringEnumerable
         {
             [Encrypt]
-            public string[] MyStrings { get; set; }
+            public IEnumerable<string> MyStrings { get; set; }
         }
 
         [Fact]
-        public void Int_array_works()
+        public void Int_enumerable_works()
         {
             Coordinator.ConfigureDefault("test");
 
             var myInts = new[] { int.MinValue, -1, 0, 1, int.MaxValue };
-            var foo = new FooIntArray { MyInts = myInts };
+            var foo = new FooIntEnumerable { MyInts = myInts };
             var json = JsonSerializer.Serialize(foo);
 
             // make sure it's encrypted
             using (var jsonDoc = JsonDocument.Parse(json))
             {
-                var jsonProperty = jsonDoc.RootElement.GetProperty(nameof(FooIntArray.MyInts));
+                var jsonProperty = jsonDoc.RootElement.GetProperty(nameof(FooIntEnumerable.MyInts));
                 jsonProperty.ValueKind.ShouldBe(JsonValueKind.Array);
 
                 var jsonProperties = jsonProperty.EnumerateArray().ToList();
@@ -67,14 +68,14 @@ namespace JsonCryption.Tests.AcceptanceTests
             }
 
             // decrypt and check
-            var decrypted = JsonSerializer.Deserialize<FooIntArray>(json);
-            decrypted.MyInts.ShouldBe(myInts);
+            var decrypted = JsonSerializer.Deserialize<FooIntEnumerable>(json);
+            decrypted.MyInts.ShouldBe(myInts);;
         }
 
-        private class FooIntArray
+        private class FooIntEnumerable
         {
             [Encrypt]
-            public int[] MyInts { get; set; }
+            public IEnumerable<int> MyInts { get; set; }
         }
     }
 }
