+++ src/JsonCryption.System.Text.Json.Tests/AcceptanceTests/KeyValuePairs.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text.Json;
 using Xunit;
 
-namespace JsonCryption.Tests.AcceptanceTests
+namespace JsonCryption.System.Text.Json.Tests.AcceptanceTests
 {
     public class KeyValuePairs
     {
