+++ src/JsonCryption.System.Text.Json.Tests/AcceptanceTests/Special_cases.cs
@@ -3,7 +3,7 @@ using System;
 using System.Text.Json;
 using Xunit;
 
-namespace JsonCryption.Tests.AcceptanceTests
+namespace JsonCryption.System.Text.Json.Tests.AcceptanceTests
 {
     public class Special_cases
     {
